more work on readme andd some  work on Emulator

+++ LibChip32/Registers.cs
@@ -5,6 +5,6 @@
         public uint[] V = new uint[16];
         public uint I { get; set; }
         public uint PC { get; set; } = 0x200;
-        public byte SP { get; set; }
+        public uint SP { get; set; }
     }
 }
